changed output to print long numbers

+++ libspmvm/src/mpihelper.c
@@ -1011,13 +1011,13 @@ LCRP_TYPE* setup_communication_parallel(CR_TYPE* cr, char *matrixPath, int optio
 
 	/* read col */
 	offset_in_file = (4+lcrp->nRows+1)*sizeof(int) + (lcrp->lfEnt[me])*sizeof(int);
-	IF_DEBUG(1) printf("PE%i: read col -- offset=%i | %d\n",me,(int)offset_in_file,lcrp->lfEnt[me]);
+	IF_DEBUG(1) printf("PE%i: read col -- offset=%lu | %d\n",me,(size_t)offset_in_file,lcrp->lfEnt[me]);
 	MPI_safecall(MPI_File_seek(file_handle, offset_in_file, MPI_SEEK_SET));
 	MPI_safecall(MPI_File_read(file_handle, lcrp->col, lcrp->lnEnts[me], MPI_INTEGER, &status));
 
 	/* read val */
 	offset_in_file = (4+lcrp->nRows+1)*sizeof(int) + (lcrp->nEnts)*sizeof(int) + (lcrp->lfEnt[me])*sizeof(data_t);
-	IF_DEBUG(1) printf("PE%i: read val -- offset=%i\n",me,(int)offset_in_file);
+	IF_DEBUG(1) printf("PE%i: read val -- offset=%lu\n",me,(size_t)offset_in_file);
 	MPI_safecall(MPI_File_seek(file_handle, offset_in_file, MPI_SEEK_SET));
 	MPI_safecall(MPI_File_read(file_handle, lcrp->val, lcrp->lnEnts[me], MPI_MYDATATYPE, &status));
